requirements.txt
Django==1.11.5
djangorestframework==3.6.4
MySQL_python==1.2.5
gunicorn==19.7.1